Firebase Manager part included
