+++ Parking Manager/Parking Manager/View Model/BaseVM.swift
@@ -1,0 +1,13 @@
+//
+//  BaseVM.swift
+//  Parking Manager
+//
+//  Created by Sanath Kumar M S on 04/09/19.
+//  Copyright © 2019 YML. All rights reserved.
+//
+
+import Foundation
+
+class BaseVM {
+    
+}
